Supermarket-Chatbot-AI-Powered-Customer-Support-Application-using-RAG-system--main/src/db_connect.py
import mysql.connector
from mysql.connector import Error
from dotenv import load_dotenv
import os

load_dotenv()

def get_db_connection():
    try:
        connection = mysql.connector.connect(
            host=os.getenv("MYSQL_HOST"),
            user=os.getenv("MYSQL_USER"),
            password=os.getenv("MYSQL_PASSWORD"),
            database=os.getenv("MYSQL_DB")
        )
        if connection.is_connected():
            return connection
    except Error as e:
        print(f"Error connecting to MySQL: {e}")
        return None
    return None

def fetch_all_products():
    connection = get_db_connection()
    if not connection:
        return []

    try:
        cursor = connection.cursor(dictionary=True)
        query = "SELECT * FROM Product_details"
        cursor.execute(query)
        products = cursor.fetchall()
        cursor.close()
        connection.close()
        return products
    except Error as e:
        print(f"Error fetching products: {e}")
        return []
    finally:
        if connection.is_connected():
            cursor.close()
            connection.close()

def fetch_product_by_id(product_id):
    connection = get_db_connection()
    if not connection:
        return None

    try:
        cursor = connection.cursor(dictionary=True)
        query = "SELECT * FROM Product_details WHERE product_id = %s"
        cursor.execute(query, (product_id,))
        product = cursor.fetchone()
        cursor.close()
        connection.close()
        return product
    except Error as e:
        print(f"Error fetching product {product_id}: {e}")
        return None
    finally:
        if connection.is_connected():
            cursor.close()
            connection.close()

