Supermarket-Chatbot-AI-Powered-Customer-Support-Application-using-RAG-system--main/requirements.txt
flask==2.3.3
streamlit==1.35.0
mysql-connector-python==8.3.0
faiss-cpu==1.7.4
sentence-transformers==2.6.1
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4
watchdog==3.0.0